% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as_canonical}
\alias{as_canonical}
\title{Canonicalize SMILES}
\usage{
as_canonical(smiles)
}
\arguments{
\item{smiles}{Character vector of SMILES.}
}
\value{
Character vector of canonical SMILES.
}
\description{
Convert SMILES to canonical form.
}
\examples{
as_canonical(c("CCO", "C(C)O"))
}
\keyword{internal}
