% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{catf}
\alias{catf}
\title{catf function}
\usage{
catf(
  ...,
  prefix = .Options$FastRet.catf.prefix,
  end = .Options$FastRet.catf.end
)
}
\arguments{
\item{...}{Arguments to be passed to sprintf for string formatting.}

\item{prefix}{A function returning a string to be used as the prefix. Default is a timestamp.}

\item{end}{A string to be used as the end of the message. Default is a newline character.}
}
\value{
No return value. This function is called for its side effect of printing a message.
}
\description{
This function prints a formatted string with optional prefix and end strings.
}
\examples{
catf("Hello, \%s!", "world")
catf("Goodbye", prefix = NULL, end = "!\n")
}
\keyword{internal}
