% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcds.R
\docType{data}
\name{CDFeatures}
\alias{CDFeatures}
\title{Chemical Descriptors}
\format{
An object of class \code{character} of length 241.
}
\usage{
CDFeatures
}
\description{
Vectors containing the feature names of the chemical descriptors listed in \link{CDNames}.
}
\seealso{
\link{CDNames}
}
\keyword{internal}
