% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcds.R
\docType{data}
\name{CDs}
\alias{CDs}
\title{Chemical Descriptors}
\format{
An object of class \code{character} of length 454.
}
\usage{
CDs
}
\description{
This object contains the feature names of various chemical descriptors.
}
\seealso{
\link{CDNames}
}
\keyword{internal}
