% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro.R
\name{check_lm_suitability}
\alias{check_lm_suitability}
\title{Checks which chemical descriptors are suitable for linear models}
\usage{
check_lm_suitability(verbose = FALSE)
}
\arguments{
\item{verbose}{A logical value indicating whether to print verbose output. Default is FALSE.}
}
\value{
A data frame with the predictors and their suitability status.
}
\description{
This function checks which chemical descriptors are suitable for use in linear model. Chemical descriptors with missing values, near-zero variance or strong outlier values are considered as not suitable. The analysis is performed using the HILIC dataset from the \href{https://www.retip.app/}{Retip} package.
}
\seealso{
\code{\link[=plot_lm_suitability]{plot_lm_suitability()}}
}
\keyword{internal}
