% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{collect}
\alias{collect}
\title{Collect elements from a list of lists}
\usage{
collect(xx)
}
\arguments{
\item{xx}{A list of lists where each inner list has the same names.}
}
\value{
A list where each element corresponds to a name of the inner list that is extracted from each inner list.
}
\description{
This function takes a list of lists where each inner list has the same names. It returns a list where each element corresponds to a name of the inner list that is extracted from each inner list. Especially useful for collecting results from lapply.
}
\examples{
xx <- lapply(1:3, function(i) list(a = i, b = i^2, c = i^3))
ret <- collect(xx)
}
\keyword{internal}
