% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{FastRet}
\alias{FastRet}
\title{Shiny app FastRet main function

This function shows you the FastRet GUI}
\usage{
FastRet()
}
\arguments{
\item{open}{Defaults to TRUE.}
}
\description{
Shiny app FastRet main function

This function shows you the FastRet GUI
}
\keyword{FastRet}
