% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastret.workflow.R
\name{fastret.workflow}
\alias{fastret.workflow}
\title{Retention time prediction  workflow}
\usage{
fastret.workflow(
  data,
  method = "glmnet",
  verbose = FALSE,
  data_set_name = "data set",
  final_model = T,
  preprocessed = F,
  interaction_terms = F,
  nfolds = 2,
  include_polynomial = F,
  degree_polynomial = 2,
  scale = T
)
}
\arguments{
\item{data}{data.frame with columns NAME, RT, SMILES}

\item{method}{prediction algorithm, either glmnet or xgboost}

\item{verbose}{additional print outputs to user if TRUE}

\item{data_set_name}{name of dataset will appear on validation plot}

\item{final_model}{TRUE if final model trained on whole dataset should be returned}

\item{preprocessed}{TRUE if data is already preprocessed and descriptor varialbes are already added}

\item{interaction_terms}{TRUE if interaction terms between all variables should be added}

\item{nfolds}{number of folds for cross validation}

\item{include_polynomial}{TRUE if polynomial terms should be added to descriptor set}

\item{degree_polynomial}{specifies degree up until which polynomials will be added if include_polynomials == TRUE}

\item{scale}{if TRUE, all variables will be centered to a mean of 0 and scaled to a standard deviation of 1}
}
\description{
Whole retention time prediction workflow. Function creates predictor set with RCDK based on SMILES.
Trains a chosen predcition model and validates the approach with a cross validation.
}
\keyword{FastRet}
