% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{fastret_app}
\alias{fastret_app}
\title{The FastRet GUI}
\usage{
fastret_app(port = 8080, host = "0.0.0.0", reload = FALSE, nsw = 1)
}
\arguments{
\item{port}{The port the application should listen on}

\item{host}{The address the application should listen on}

\item{reload}{Whether to reload the application when the source code changes}

\item{nsw}{The number of subworkers each worker is allowed to start. The higher this
number, the faster individual tasks like model fitting can be processed.}
}
\value{
An object of class \code{shiny.appobj}.
}
\description{
Creates the FastRet GUI
}
\examples{
x <- fastret_app()
if (interactive()) shiny::runApp(x)
}
\keyword{public}
