% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get cache directory}
\usage{
get_cache_dir(subdir = NULL)
}
\arguments{
\item{subdir}{Optional subdirectory within the cache directory.}
}
\value{
The path to the cache directory or subdirectory.
}
\description{
Creates and returns the cache directory for the FastRet package.
}
\examples{
path <- get_cache_dir()
}
\keyword{internal}
