% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_predictors}
\alias{get_predictors}
\title{Extract predictor names from an 'frm' object}
\usage{
get_predictors(frm, base = TRUE, adjust = FALSE)
}
\arguments{
\item{frm}{An object of class 'frm' from which to extract the predictor names.}

\item{base}{Logical indicating whether to include base model predictors.}

\item{adjust}{Logical indicating whether to include adjustment model predictors.}
}
\value{
A character vector with the predictor names.
}
\description{
Extracts the predictor names from an 'frm' object.
}
\examples{
frm <- read_rp_lasso_model_rds()
get_predictors(frm)
}
\keyword{internal}
