% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HILIC}
\alias{HILIC}
\title{Retention Times (RT) measured on a Hydrophilic Interaction Chromatography (HILIC) Column}
\format{
A dataframe of 393 metabolites with the following columns:
\describe{
\item{NAME}{Name of the metabolite}
\item{SMILES}{SMILES notation of the metabolite}
\item{RT}{Retention time}
}
}
\source{
Measured by the Institute of Functional Genomics at the University of
Regensburg.
}
\usage{
HILIC
}
\description{
Retention time data from a hydrophilic interaction chromatography experiment.
}
\keyword{dataset}
