% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{init_log_dir}
\alias{init_log_dir}
\title{Initialize log directory}
\usage{
init_log_dir(SE)
}
\arguments{
\item{SE}{A list containing session information.}
}
\value{
Updates the logdir element in the SE list with the path to the log directory.
}
\description{
This function initializes the log directory for the session. It creates a new directory if it does not exist.
}
\examples{
SE <- as.environment(list(session = list(token = "asdf")))
init_log_dir(SE)
dir.exists(SE$logdir)
}
\keyword{internal}
