% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{now}
\alias{now}
\title{now function}
\usage{
now(format = "\%Y-\%m-\%d \%H:\%M:\%OS2")
}
\arguments{
\item{format}{A string representing the desired time format. Default is "\%Y-\%m-\%d \%H:\%M:\%OS2".}
}
\value{
A string representing the current system time in the specified format.
}
\description{
This function returns the current system time formatted according to the provided format string.
}
\examples{
now()            # e.g. "2024-06-12 16:09:32.41"
now("\%H:\%M:\%S")  # e.g. "16:09:32"
}
\keyword{internal}
