% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro.R
\name{plot_lm_suitability}
\alias{plot_lm_suitability}
\title{Plot the suitability of predictors for linear models}
\usage{
plot_lm_suitability(
  slist = check_lm_suitability(),
  pdfpath = tempfile("lm_suitability", fileext = ".pdf")
)
}
\arguments{
\item{slist}{A list containing the data frame \code{df}, the matrix \code{X}, and the data frame \code{V} from \code{check_lm_suitability()}.}

\item{pdfpath}{The path to the pdf file to save the plots.}
}
\value{
No return value. The function is used for its side effect of creating a pdf file with the plots.
}
\description{
This function creates one pdf page for every predictor inside \code{slist$X}. The pdf page consists of the following three plots shown next to each other:
\enumerate{
\item Histogram
\item Density plot
\item Scatterplot against \code{slist$df$RT}
The name of the predictor, its suitability, and the status of the checks for missing values, near-zero variance, and outliers are shown in the title of each plot.
}
}
\examples{
\donttest{
slist <- check_lm_suitability()
plot_lm_suitability(slist)
}
}
\seealso{
\code{\link[=check_lm_suitability]{check_lm_suitability()}}
}
\keyword{internal}
