% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{ram_cache}
\alias{ram_cache}
\title{RAM Cache Environment}
\format{
An environment with the following elements:
\itemize{
\item \code{CDs}: A named list where each element contains the chemical descriptors
for one SMILES string. The names of the list are SMILES strings, and each
element is a data.frame with one row containing the chemical descriptors
listed in \link{CDFeatures}.
}
}
\usage{
ram_cache
}
\description{
An environment used for caching data in RAM.
}
\details{
This environment is used by \code{\link[=getCDsFor1Molecule]{getCDsFor1Molecule()}} to store the results of
previous calculations to speed up subsequent calls. It gets initalized upon
the first call of \code{\link[=getCDsFor1Molecule]{getCDsFor1Molecule()}} with the chemical descriptors for
all molecules available in the \link{RP} dataset and the \code{HILIC} dataset of the
\href{https://www.retip.app/}{Retip} package.
}
\examples{
dim(ram_cache$CDs) # 1316 241
}
\references{
Retip: Retention Time Prediction for Compound Annotation in Untargeted
Metabolomics Paolo Bonini, Tobias Kind, Hiroshi Tsugawa, Dinesh Kumar
Barupal, and Oliver Fiehn Analytical Chemistry 2020 92 (11), 7515-7522 DOI:
10.1021/acs.analchem.9b05765
}
\keyword{internal}
