% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_retip_hilic_data}
\alias{read_retip_hilic_data}
\title{Download and read the HILIC dataset from Retip the package}
\usage{
read_retip_hilic_data(verbose = 1)
}
\arguments{
\item{verbose}{Verbosity level. 1 == print progress messages, 0 == no progress messages.}
}
\value{
df A data frame containing the HILIC dataset.
}
\description{
Downloads and reads the HILIC dataset from the \href{https://www.retip.app/}{Retip} package. The dataset is downloaded from \verb{https://github.com/oloBion/Retip/raw/master/data/HILIC.RData}, saved to a temporary file and then read and returned.
}
\examples{
df <- read_retip_hilic_data(verbose = 0)
}
\references{
Retip: Retention Time Prediction for Compound Annotation in Untargeted Metabolomics Paolo Bonini, Tobias Kind, Hiroshi Tsugawa, Dinesh Kumar Barupal, and Oliver Fiehn Analytical Chemistry 2020 92 (11), 7515-7522 DOI: 10.1021/acs.analchem.9b05765
}
\keyword{dataset}
