% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_rp_lasso_model_rds}
\alias{read_rp_lasso_model_rds}
\title{LASSO Model trained on RP dataset}
\usage{
read_rp_lasso_model_rds()
}
\value{
A \code{frm} object.
}
\description{
Read a LASSO model trained on the \link{RP} dataset using \code{\link[=train_frm]{train_frm()}}.
}
\examples{
frm <- read_rp_lasso_model_rds()
}
\keyword{dataset}
