% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RP_AXMM}
\alias{RP_AXMM}
\title{Retention Times (RT) measured on a Reverse Phase Anion Exchange Mixed Mode
(RP-AXMM) Column}
\format{
A dataframe of 438 metabolites with the following
columns:
\describe{
\item{SMILES}{SMILES notation of the metabolite}
\item{NAME}{Name of the metabolite}
\item{RT}{Retention time}
}
}
\source{
Measured by the Institute of Functional Genomics at the University of
Regensburg.
}
\usage{
RP_AXMM
}
\description{
Retention time data from a Reverse Phase Anion Exchange Mixed Mode (RP-AXMM)
Column.
}
\seealso{
read_rp_xlsx
}
\keyword{dataset}
