% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.sm.R
\name{shiny.sm}
\alias{shiny.sm}
\title{Prepare data by scaling it
Train Ridge Regression on all data}
\usage{
shiny.sm(raw_data, method, k_cluster)
}
\description{
Use the coefficients of trained model as feature weights
Apply Selective Measuring 1.0 on weighted features
}
