% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{withStopMessage}
\alias{withStopMessage}
\title{Execute an expression and print an error message if it fails}
\usage{
withStopMessage(expr)
}
\arguments{
\item{expr}{The expression to execute}
}
\value{
The result of the expression
}
\description{
Execute an expression and print an error message if it fails
}
\keyword{internal}
