% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{withTimeout}
\alias{withTimeout}
\title{Execute an expression with a timeout}
\usage{
withTimeout(expr, timeout = 2)
}
\arguments{
\item{expr}{The expression to execute}

\item{timeout}{The timeout in seconds. Default is 2.}
}
\value{
The result of the expression
}
\description{
Execute an expression with a timeout
}
\examples{
withTimeout(cat("This works\n"), timeout = 0.2)
try(withTimeout({Sys.sleep(0.2); cat("This will fail\n")}, timeout = 0.1))
}
\keyword{internal}
