% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcds.R
\name{analyzeCDNames}
\alias{analyzeCDNames}
\title{Analyze Chemical Descriptors Names}
\usage{
analyzeCDNames(
  df = read_rp_xlsx(),
  descriptors = rcdk::get.desc.names(type = "all")
)
}
\arguments{
\item{df}{dataframe with two mandatory columns: "NAME" and "SMILES"}

\item{descriptors}{vector of chemical descriptor names}
}
\description{
Analyze the chemical descriptor names and return a dataframe with their names and a boolean column indicating if all values are NA.
}
\details{
This function is used to analyze the chemical descriptor names and to identify which descriptors produce only NAs in the test datasets. The function is used to generate the CDNames object.
}
\examples{
df <- head(read_rp_xlsx(), 3)
descriptors <- head(rcdk::get.desc.names(type = "all"))
X <- analyzeCDNames(df, descriptors)
}
\keyword{internal}
