% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcds.R
\name{analyzeCDNames}
\alias{analyzeCDNames}
\title{Analyze Chemical Descriptors Names}
\usage{
analyzeCDNames()
}
\description{
Analyze the chemical descriptors names and return a dataframe with the names and a boolean column indicating if all values are NA.
}
\details{
This function is used to analyze the chemical descriptors names and to identify which descriptors produce only NAs in the test datasets. The function is used to generate the CDNames object.
}
\keyword{internal}
