% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_predictors}
\alias{get_predictors}
\title{Extract predictor names from an 'frm' object}
\usage{
get_predictors(frm = train_frm())
}
\arguments{
\item{frm}{An object of class 'frm' from which to extract the predictor names.}
}
\value{
A character vector with the predictor names.
}
\description{
This function extracts the predictor names from an 'frm' object.
}
\examples{
\donttest{
frm <- train_frm()
get_predictors(frm)
}
}
\keyword{internal}
