% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCD.R
\name{getCD}
\alias{getCD}
\title{getCD}
\usage{
getCD(x, verbose = FALSE)
}
\arguments{
\item{x}{dataframe with two mandatory column: "Name" and "SMILES"}

\item{verbose}{TRUE if additional print output should be shown}
}
\description{
Calculate Chemical Descriptors using RCDK, function is based on Retip::getCD.
}
\keyword{FastRet}
