% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcds.R
\name{getCDs}
\alias{getCDs}
\title{Get Chemical Descriptors for a list of molecules}
\usage{
getCDs(df = read_rp_xlsx(), verbose = 1, nw = 1)
}
\arguments{
\item{df}{dataframe with two mandatory columns: "NAME" and "SMILES"}

\item{verbose}{0: no output, 1: progress, 2: more progress and warnings}

\item{nw}{number of workers for parallel processing}
}
\description{
Calculate Chemical Descriptors for a list of molecules. Molecules can appear multiple times in the list.
}
\keyword{public}
