% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_retip_hilic_data}
\alias{read_retip_hilic_data}
\title{Read the HILIC dataset from the Retip package}
\usage{
read_retip_hilic_data(verbose = 1)
}
\arguments{
\item{verbose}{Verbosity level. Set to 1 to enable progress messages,
0 to suppress them.}
}
\value{
A data frame containing the Retip HILIC dataset.
}
\description{
If the \href{https://www.retip.app/}{Retip} package is installed, its HILIC
dataset is read from its installation directory. If not, the dataset is
downloaded from
\verb{https://github.com/oloBion/Retip/raw/master/data/HILIC.RData} to a temporary
file and then read from there.
}
\examples{
df <- read_retip_hilic_data(verbose = 0)

}
\references{
Retip: Retention Time Prediction for Compound Annotation in Untargeted
Metabolomics Paolo Bonini, Tobias Kind, Hiroshi Tsugawa, Dinesh Kumar
Barupal, and Oliver Fiehn Analytical Chemistry 2020 92 (11), 7515-7522 DOI:
10.1021/acs.analchem.9b05765
}
\keyword{dataset}
