% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP.R
\docType{data}
\name{RP}
\alias{RP}
\title{Retention time data from a reverse phase liquid chromatography.}
\format{
A tibble of size 442 metabolites and their retention times.
\describe{
\item{RT}{Retention time in minutes}
\item{SMILES}{SMILES of Metabolite}
\item{NAME}{Name of Metabolite}
}
}
\source{
measured by functional genomics lab at the university of Regensburg
}
\usage{
data(RP)
}
\description{
Measured with a temperature of 35 degree and a flowrate of 0.3ml/min
}
\keyword{datasets}
