% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RP_Mod}
\alias{RP_Mod}
\title{Subset of the 'RP' dataset measured under different conditions}
\format{
A named list with elements: 'Normal', 'Steep', 'Flat', 'T25_Flat',
'FR25_Flat', 'T25_FR25_Flat', and 'T25_FR25_Steep'. Each element is a data
frame with 25 metabolites (rows) and the following columns:
\describe{
\item{RT}{Retention time}
\item{SMILES}{SMILES notation of the metabolite}
\item{NAME}{Name of the metabolite}
}
The 'Normal' condition is a true subset of the original \link{RP} dataset: it
contains the same metabolites with identical retention times as in \link{RP}. All
other elements ('Steep', 'Flat', etc.) contain the same metabolites, but
their retention times were measured under different chromatographic
conditions.
}
\source{
Measured by the Institute of Functional Genomics at the University of
Regensburg.
}
\usage{
RP_Mod
}
\description{
A subset of 25 metabolites from the original \link{RP} dataset was re-measured
under modified chromatographic conditions (gradient, temperature, and/or flow
rate). This object contains the results as data frames. See 'Details' for an
overview of the modifications.
}
\details{
The following conditions were used during the experiment:\tabular{llll}{
   Name \tab Temperature \tab Flow Rate \tab Gradient \cr
   Normal \tab 35 \tab 0.3 \tab normal \cr
   Steep \tab 35 \tab 0.3 \tab steep \cr
   Flat \tab 35 \tab 0.3 \tab flat \cr
   T25_Flat \tab 25 \tab 0.3 \tab flat \cr
   FR025_Flat \tab 35 \tab 0.25 \tab flat \cr
   T25_FR025_Flat \tab 25 \tab 0.25 \tab flat \cr
   T25_FR025_Steep \tab 25 \tab 0.25 \tab steep \cr
}


For all conditions, formic acid 0.1\% (v/v) in Water was used as mobile phase
A and formic acid 0.1\% (v/v) in ACN as mobile phase B. Temperatures are given
in \eqn{^\circ}C and flow rates in mL/min. For the exact definitions of a
'normal', 'steep' and 'flat' gradient please refer to the FastRet
publication.
}
\keyword{dataset}
