% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{selective_measuring}
\alias{selective_measuring}
\title{Selective Measuring Function}
\usage{
selective_measuring(raw_data = read_rp_xlsx(), k_cluster = 25, verbose = 1)
}
\arguments{
\item{raw_data}{The raw data to be processed. Default is the result of \code{read_rp_xlsx()}.}

\item{k_cluster}{The number of clusters for PAM clustering. Default is 25.}

\item{verbose}{The level of verbosity. Default is 1.}
}
\value{
A list containing the following elements:
\itemize{
\item clustering: a data frame with raw data, cluster assignments, and medoid indicators
\item clobj: the PAM clustering object
\item coefs: the coefficients from the Ridge Regression model
\item model: the Ridge Regression model
\item df: the preprocessed data
\item dfz: the standardized features
\item dfzb: the features scaled by coefficients of the Ridge Regression model
}
}
\description{
This function performs selective measuring on the provided raw data. It includes preprocessing, standardizing features, training a Ridge Regression model, scaling features by coefficients of the Ridge Regression model, applying PAM clustering, and returning clustering results.
}
\keyword{public}
