% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{withLineEnd}
\alias{withLineEnd}
\title{Add line end}
\usage{
withLineEnd(x)
}
\arguments{
\item{x}{A string.}
}
\value{
The input string with a newline character at the end if it was not already present.
}
\description{
This function checks if a string ends with a newline character. If not, it adds one.
}
\examples{
cat(withLineEnd("Hello"))
}
\keyword{internal}
